Remove move "enabled" from Reoseta::Config to Rosetta class
