+++ lib/rosetta.rb
@@ -14,8 +14,7 @@ require 'rosetta/repositories/onesky'
 module Rosetta
   class << self
 
-    delegate :repository, :setup, :enabled?, :enable, :disable,
-             to: :config
+    delegate :repository, :setup, to: :config
 
     def config
       @config ||= Config.new
@@ -30,11 +29,37 @@ module Rosetta
     end
 
     def phrases
-      RequestStore.store[:used_phrases] ||= []
+      request[:used_phrases] ||= []
     end
 
     def locale
       I18n.locale
+    end
+
+    def enabled?
+      active
+    end
+
+    def enable
+      self.active = true
+    end
+
+    def disable
+      self.active = false
+    end
+
+    private
+
+    def active
+      request[:actived].nil? ? false : request[:actived]
+    end
+
+    def active=(value)
+      request[:actived] = value
+    end
+
+    def request
+      RequestStore.store[:rosetta] ||= {}
     end
 
   end
