+++ lib/rosetta/config.rb
@@ -7,45 +7,22 @@ module Rosetta
       set_defaults!
     end
 
-    def onesky
-      @onesky ||= OneSky.new
-    end
-
     def set_defaults!
       @repository = Rosetta::Repositories::Local
-      disable
-    end
-
-    class OneSky
-
-      attr_accessor :project_id, :subdomain
-
     end
 
     def setup
       yield self
     end
 
-    def enabled?
-      active
+    def onesky
+      @onesky ||= OneSky.new
     end
 
-    def enable
-      self.active = true
-    end
+    class OneSky
 
-    def disable
-      self.active = false
-    end
+      attr_accessor :project_id, :subdomain
 
-    private
-
-    def active
-      RequestStore.store[:rosetta_enabled] ||= false
-    end
-
-    def active=(value)
-      RequestStore.store[:rosetta_enabled] = value
     end
 
   end
