+++ lib/rosetta/phrase.rb
@@ -1,15 +1,17 @@
-class Phrase
+module Rosetta
+  class Phrase
 
-  attr_reader :keys, :text, :repository_link
+    attr_reader :keys, :text, :repository_link
 
-  def initialize(keys, text, repository_link)
-    @keys = keys
-    @text = text
-    @repository_link = repository_link
+    def initialize(keys, text, repository_link)
+      @keys = keys
+      @text = text
+      @repository_link = repository_link
+    end
+
+    def code
+      @keys.join('.')
+    end
+
   end
-
-  def code
-    @keys.join('.')
-  end
-
 end
