+++ spec/rosetta/features/local_repository_spec.rb
