+++ spec/rosetta/i18n_with_stored_phrases_spec.rb
