+++ spec/rosetta/repository/local_spec.rb
