+++ spec/rosetta/repository/onesky_spec.rb
