+++ spec/rosetta/rosetta_spec.rb
